# contract/openapi.yaml
...
openapi: 3.0.3
info:
  title: T-Shirt Store API
  version: 0.1.0
  description: |
    Contract for the T-Shirt Store API. Week 3 implements this document. If the
    implementation and the document disagree, the document is not out of date.

    Written against the schema in `4-database/3-erd/store.dbml`, which this branch
    carries since the ERD merge at `f96d62a`.

    **Money.** Every amount is an integer in minor units. 1999 means 19.99. The store
    is single-currency and prices are in USD. No response carries a currency field.

    **Tokens.** An access token is valid for 15 minutes. A refresh token is valid for
    7 days and rotates on every use. Sign-out deletes the presented device's refresh
    row. It does not revoke an access token the server already issued, so revocation
    lags by up to 15 minutes. Changing or resetting a password deletes every refresh row for
    that user.

    **Rate limits.** The server limits every operation by source address, in three tiers.
    Browsing carries the loosest limit. Sign-in carries a tighter limit. The three password
    operations carry the tightest limit. The server answers 429 with a `Retry-After` header
    when it refuses a request. This document declares the 429 response only where a client
    must handle it, but the server can answer 429 on any path.
servers:
  - url: http://localhost:3000/v1
    description: Local development. The /v1 prefix lives here, not on each path key.
...
security:
  - bearerAuth: []
paths:
...
  /auth/sessions:
    post:
      tags: [auth]
      summary: Create a session and get tokens
      description: |
        This operation is public. It does not need a token.

        Send the email and the password. The server creates a session for this device.
        The server returns an access token and a refresh token.

        A wrong email or a wrong password returns 401 with the `invalid-credentials`
        problem type. The server does not say which of the two was wrong.
      operationId: createSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                # Every property below carries the rule its 400 fires on. `Problem.errors`
                # promises one message per rejected field, so a field with no rule is a
                # message the server can never produce. The ERD declares all 18 of its
                # varchars without a length, so these bounds are the contract's own and the
                # schema has to be given the matching ones.
                email:
                  type: string
                  format: email
                  maxLength: 254
                  description: RFC 5321 4.5.3.1.3 caps a forward-path at 256 octets
                    including the punctuation, which leaves 254 for the address.
                  example: ana@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  maxLength: 128
                  description: The maximum is a cost bound, not a security rule. The server
                    hashes the value it receives. An unbounded field lets one request
                    consume a large amount of processor time.
                deviceName:
                  type: string
                  maxLength: 64
                  description: A label for this device. The user sees this label in the
                    device list.
                  example: Ana iPhone
      responses:
        '201':
          description: The server created the session.
          headers:
            Location:
              description: The URL of the new session. Send a DELETE to that URL to sign
                this device out.
              schema:
                type: string
              example: /v1/auth/sessions/42
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionTokens'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: [auth]
      summary: List the devices signed in to this account
      description: |
        The server returns one entry for each device that is signed in to this account.
        The server returns the entries in pages. Use `limit` and `offset` to select a page.
      operationId: listSessions
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: The devices signed in to this account.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/sessions/current:
    delete:
      tags: [auth]
      summary: Sign out this device
      description: |
        The server ends the session of the device that sends this request, and deletes
        every refresh token in it. Other devices stay signed in.

        The access token stops working at once. The server checks on every request that
        the session named by the token still exists, so a device that signs out cannot
        keep acting until its token expires.
      operationId: deleteCurrentSession
      responses:
        '204':
          description: The server deleted the session.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/sessions/{id}:
    delete:
      tags: [auth]
      summary: Sign out another device
      description: |
        The server ends one session of this user, and deletes every refresh token in it.
        The access token of that device stops working at once.

        A session id that belongs to another user returns 404. A 403 would confirm that
        the session exists.
      operationId: deleteSession
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: The server deleted the session.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Exchange a refresh token for new tokens
      description: |
        This operation is public. It does not need an access token.

        Send the refresh token. The server returns a new access token and a new refresh
        token. The old refresh token stops working.

        The server keeps one session for each device. A session can hold more than one live
        refresh token for a few seconds after a rotation, and `GET /auth/sessions` still
        reports it as one device. The session id does not change, so an id from
        `GET /auth/sessions` stays valid for the life of the device session, and
        `DELETE /auth/sessions/{id}` ends every token in that session.

        A refresh token works one time only. If the server receives one that was already
        used, the server deletes every refresh token for that user. The server assumes an
        attacker stole the token. Each device must sign in again.

        The server allows one exception, for a few seconds after a rotation. If the client
        sends a token the server has already spent, and it was spent inside that window, the
        server issues a new token in the same session. It deletes nothing, and it does not
        stop the token it issued to the first caller. Two tabs of one application can send
        the same token at the same moment, and the server must not read that as theft. The
        window belongs to the spent token, so it does not extend when that token is sent
        again.

        A token the server never issued is not a replay. The server answers 401 and deletes
        nothing.

        A refresh token that is unknown, expired or already used returns 401 with the
        `refresh-token-unknown` problem type.

        The server limits the request rate. Too many requests return 429.
      operationId: refreshSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  minLength: 1
                  maxLength: 512
                  description: The value from the previous call. It is opaque, so the
                    only rules the contract can state are that it is present and bounded.
      responses:
        '200':
          description: The server issued new tokens.
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/SessionTokens'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/forgot-password:
    post:
      tags: [auth]
      summary: Request a password reset link
      description: |
        This operation is public. It does not need a token.

        Send the email address. The server sends a reset link to that address.

        The server answers 202 whether or not the address has an account. A different
        answer would tell the caller which addresses are registered.

        The server limits the request rate. Too many requests return 429.
      operationId: requestPasswordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
                  example: ana@example.com
      responses:
        '202':
          description: The server accepted the request. It sends the mail only if the
            address has an account.
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/reset-password:
    post:
      tags: [auth]
      summary: Set a new password with a reset token
      description: |
        This operation is public. It does not need a token.

        Send the token from the reset mail and the new password. The server sets the new
        password.

        The token travels in the body and never in the path. A path segment reaches the
        access log and the `Referer` header.

        A token that is unknown or expired returns 422. The body is well formed and the
        server rejects it on its content.

        On success the server deletes every refresh row for this user. Each device must
        sign in again. The server also sends an email to the account address.
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                  minLength: 1
                  maxLength: 512
                  description: The value from the reset mail. It is opaque and it is
                    valid one time only.
                password:
                  type: string
                  format: password
                  minLength: 8
                  maxLength: 128
                  description: The new password. The same bounds apply as in `createUser`.
      responses:
        '204':
          description: The server set the new password and ended every session.
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users:
    post:
      tags: [auth]
      summary: Create an account
      description: |
        This operation is public. It does not need a token.

        Send an email address, a password and a name. The server creates a client
        account. The server returns the new account and a `Location` header that names
        it.

        The request does not accept a role. Every account this operation creates is a
        client account. This contract has no operation that changes a role, so a manager
        account is seeded.

        An email address that is already registered returns 409 with the `email-taken`
        problem type.

        The server limits the request rate. Too many requests return 429.
      operationId: createUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, firstName, lastName]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
                  description: RFC 5321 4.5.3.1.3 caps a forward-path at 256 octets
                    including the punctuation, which leaves 254 for the address.
                  example: ana@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  maxLength: 128
                  description: The maximum is a cost bound, not a security rule. The
                    server hashes the value it receives. An unbounded field lets one
                    request consume a large amount of processor time.
                firstName:
                  type: string
                  maxLength: 100
                  example: Ana
                lastName:
                  type: string
                  maxLength: 100
                  example: Ramirez
      responses:
        '201':
          description: The server created the account.
          headers:
            Location:
              description: The URL of the new account.
              schema:
                type: string
              example: /v1/users/128
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Session:
      type: object
      required: [id, createdAt, expiresAt]
      properties:
        id:
          type: integer
          description: The id of this device session. Token rotation updates the row and
            keeps the id, so a client can hold this value and send it to
            `DELETE /auth/sessions/{id}`.
          example: 42
        deviceName:
          type: string
          example: Ana iPhone
        createdAt:
          type: string
          format: date-time
          description: When this device first signed in. Token rotation does not change
            this value.
          example: '2026-08-21T09:14:00Z'
        expiresAt:
          type: string
          format: date-time
          description: When the refresh token expires. Every rotation moves this value.
          example: '2026-08-28T09:14:00Z'

    # ---- Errors, DECISIONS.md item 2 ----------------------------------------
    # RFC 9457, which obsoletes RFC 7807. Served as application/problem+json, never
    # application/json.

    # The closed set of problem kinds this API emits. A client branches on this value and
    # on nothing else. Only failures the status code cannot tell apart get a member here.
    # Everything else omits `type`, which RFC 9457 defines as `about:blank`, meaning no
    # semantics beyond the status code.
    #
    # These URIs are identifiers first. RFC 9457 says consumers SHOULD NOT dereference them
    # automatically. Week 3 serves a short page per member at /problems/{slug}.
...
    SessionTokens:
      type: object
      # Review round 5, 2026-08-25. The mentor asked that the sign-in response carry the
      # user as well as the tokens, so a client does not need a second request to learn
      # who it just signed in as. `user` is required on this response: the server has the
      # row in hand at that moment, so making it optional would only invite a client to
      # handle an absence that never happens.
      required: [accessToken, refreshToken, user]
      properties:
        user:
          allOf:
            - $ref: '#/components/schemas/User'
          description: The account that just signed in. The same shape `POST /users`
            returns, so a client parses one schema and not two.
        accessToken:
          type: string
          description: A bearer token. It is valid for 15 minutes.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.PAYLOAD.SIGNATURE
        refreshToken:
          type: string
          description: It is valid for 7 days. It rotates on every use. Keep the new
            value. Discard the old value.
          example: 8f14e45fceea167a5a36dedd4bea2543f8b3c9e1d0a7b264

    # ---- Pagination, DECISIONS.md item 3 ------------------------------------
    # Every collection returns this envelope. No endpoint returns a bare array.
    PageMeta:
      type: object
      required: [total, limit, offset]
      properties:
        total:
          type: integer
          description: The number of rows that match the filter, before limit and offset
            apply.
          example: 347
        limit:
          type: integer
          description: The number of rows in this page. The `limit` query parameter sets
            it and carries the default. The maximum is 100 in both directions.
          maximum: 100
          example: 20
        offset:
          type: integer
          description: The number of rows the server skipped. Offset pagination drifts
            if another request inserts rows during a scan.
          example: 40

    # ---- Users --------------------------------------------------------------
    # The field list is the point. `users` also carries `password_hash`,
    # `reset_token` and `reset_token_expires_at`. A Week 3 serializer that returns the row
    # whole would put a live password-reset token in a 200. Naming the fields here makes
    # that a contract violation rather than an accident.
    User:
      type: object
      required: [id, email, firstName, lastName, role, createdAt]
      properties:
        id:
          type: integer
          example: 128
        email:
          type: string
          format: email
          example: ana@example.com
        firstName:
          type: string
          example: Ana
        lastName:
          type: string
          example: Ramirez
        role:
          $ref: '#/components/schemas/Role'
        createdAt:
          type: string
          format: date-time
          example: '2026-08-21T13:45:00Z'

    # ---- Catalog ------------------------------------------------------------
    # Two product shapes on purpose. The list entry carries a price summary and no
    # variants, so one page costs a fixed number of queries however many sizes a product
    # has. The detail carries the variants, and a client asks for it one product at a
    # time. That split is the answer to the N+1 reading.
...
    Role:
      type: string
      enum: [manager, client, delivery_person]

  # ---- Query parameters, DECISIONS.md item 3 ------------------------------
  # Every collection takes this pair. The bounds repeat PageMeta, so a caller reads the
  # same numbers on the request side and on the response side. A collection that does
  # not reference both returns an unbounded array.
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: The maximum number of rows to return. The default is 20 and the
        maximum is 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    Offset:
      name: offset
      in: query
      required: false
      description: The number of rows to skip before the first row of this page.
      schema:
        type: integer
        minimum: 0
        default: 0
      example: 40
...
  responses:
...
    BadRequest:
      description: >-
        The server cannot read the request. Either a body field failed validation, and
        then `errors` names each rejected field, or a path segment that must be an
        integer is not one. A path id that is a well formed integer but matches no row
        returns 404, not 400.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Validation failed
            status: 400
            detail: One or more fields did not pass validation.
            errors:
              - field: email
                message: must be a valid email address
    Unauthorized:
      description: No token, an expired token, or a refresh row that no longer exists.
      headers:
        WWW-Authenticate:
          schema:
            type: string
          description: Required on every 401.
          example: Bearer realm="tshirt-store"
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          # One named example per `ProblemType` member this status can carry. A client
          # that cannot separate these three loops between refreshing and sending the
          # user back to the login screen. DECISIONS.md item 2.
          examples:
            expired:
              summary: The access token expired
              value:
                type: https://tshirt.store/problems/access-token-expired
                title: Access token expired
                status: 401
                detail: Refresh the token and send the request again.
            credentials:
              summary: Wrong email or wrong password
              value:
                type: https://tshirt.store/problems/invalid-credentials
                title: Invalid credentials
                status: 401
                detail: The server did not accept this email and password.
            refresh:
              summary: The refresh token is unknown or already used
              value:
                type: https://tshirt.store/problems/refresh-token-unknown
                title: Refresh token unknown
                status: 401
                detail: Sign in again. The server ended every session for this user.
...
    NotFound:
      description: The resource does not exist, or this caller may not know that it
        exists. A 403 would confirm the row exists, so an ownership failure returns 404.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          # Two causes, one status code, and the split is deliberate. DECISIONS.md
          # item 7: 403 protects an action, 404 protects a fact.
          examples:
            absent:
              summary: The resource does not exist
              value:
                title: Not found
                status: 404
                detail: The server did not find this resource.
            notYours:
              summary: It exists and belongs to another client
              value:
                title: Not found
                status: 404
                detail: The server did not find this resource.
    Conflict:
      description: The request conflicts with the current state. The same request can
        succeed later.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          # The three 409 members. A client that cannot separate them shows the wrong
          # message. DECISIONS.md item 2.
          examples:
            emailTaken:
              summary: Sign-up against an existing account
              value:
                type: https://tshirt.store/problems/email-taken
                title: Email already registered
                status: 409
                detail: An account with this email already exists.
            insufficientStock:
              summary: Fewer units on hand than the request asks for
              value:
                type: https://tshirt.store/problems/insufficient-stock
                title: Not enough stock
                status: 409
                detail: This variant has 3 units on hand and the request asks for 5.
            notCancellable:
              summary: Cancel after the order shipped
              value:
                type: https://tshirt.store/problems/order-not-cancellable
                title: Order cannot be cancelled
                status: 409
                detail: This order already shipped.
    UnprocessableEntity:
      description: The request is well formed, and invalid whatever the state. The same
        request cannot succeed later.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            resetToken:
              summary: The reset token is unknown or expired
              value:
                title: Unprocessable content
                status: 422
                detail: The reset token is unknown or expired.
            missingCategory:
              summary: A categoryIds entry names no category
              value:
                title: Unprocessable content
                status: 422
                detail: The request names a category that does not exist.
    # RFC 9110 15.5.14 names this code Content Too Large. The assigned HTTP Status Codes
    # reading prints the RFC 2616 name, Request Entity Too Large. RFC 2616 is obsolete
    # twice over, so the current name is the one used here.
...
    TooManyRequests:
      description: The caller reached the rate limit.
      headers:
        Retry-After:
          schema:
            type: string
          description: How long the caller must wait. RFC 9110 10.2.3 allows a number of
            seconds or an HTTP-date.
          example: '60'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Too many requests
            status: 429
            detail: Wait before you send this request again.
    InternalServerError:
      description: The server failed. The caller may retry the same request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Internal server error
            status: 500
            detail: The server failed to handle the request.